Moved SVOPathFindingRenderingComponent inside SVOPathFinderTest
